Refactor: move ResultItem from core data to core domain
