+++ core/database/build.gradle.kts
@@ -8,7 +8,7 @@ plugins {
 android.namespace = "com.waffiq.bazz_movies.core.database"
 
 dependencies {
-  api(project(":core:data"))
+  api(project(":core:domain"))
   api(project(":core:common"))
   implementation(project(":core:utils"))
 }
