+++ core/database/src/main/kotlin/com/waffiq/bazz_movies/core/database/utils/DatabaseMapper.kt
@@ -1,9 +1,9 @@
 package com.waffiq.bazz_movies.core.database.utils
 
+import com.waffiq.bazz_movies.core.common.utils.Constants.NOT_AVAILABLE
 import com.waffiq.bazz_movies.core.database.data.model.Favorite
 import com.waffiq.bazz_movies.core.database.data.model.FavoriteEntity
-import com.waffiq.bazz_movies.core.data.ResultItem
-import com.waffiq.bazz_movies.core.common.utils.Constants.NOT_AVAILABLE
+import com.waffiq.bazz_movies.core.domain.ResultItem
 import com.waffiq.bazz_movies.core.utils.GenreHelper.transformListGenreIdsToJoinName
 
 object DatabaseMapper {
