+++ core/domain/build.gradle.kts
@@ -4,4 +4,7 @@ plugins {
 }
 
 android.namespace = "com.waffiq.bazz_movies.core.domain"
-dependencies { implementation(libs.androidx.core.ktx) }
+dependencies {
+  implementation(project(":core:common"))
+  implementation(libs.androidx.core.ktx)
+}
