+++ core/domain/src/main/kotlin/com/waffiq/bazz_movies/core/domain/ResultItem.kt
@@ -1,4 +1,4 @@
-package com.waffiq.bazz_movies.core.data
+package com.waffiq.bazz_movies.core.domain
 
 import android.os.Parcelable
 import com.waffiq.bazz_movies.core.common.utils.Constants.MOVIE_MEDIA_TYPE
