+++ core/movie/build.gradle.kts
@@ -9,10 +9,9 @@ android.namespace = "com.waffiq.bazz_movies.core.movie"
 
 dependencies {
   api(project(":core:common"))
-  api(project(":core:data"))
-  api(project(":core:network"))
   api(project(":core:database"))
   api(project(":core:domain"))
+  api(project(":core:network"))
   implementation(project(":core:mappers"))
 
   implementation(libs.androidx.core.ktx)
