+++ core/uihelper/build.gradle.kts
@@ -5,8 +5,8 @@ plugins {
 android.namespace = "com.waffiq.bazz_movies.core.uihelper"
 
 dependencies {
-  api(project(":core:designsystem"))
   api(project(":core:common"))
+  api(project(":core:designsystem"))
 
   api(libs.androidx.core.ktx)
   implementation(libs.androidx.paging.runtime)
