+++ feature/detail/src/main/kotlin/com/waffiq/bazz_movies/feature/detail/domain/usecase/getDetailMovie/GetDetailMovieInteractor.kt
@@ -1,6 +1,7 @@
 package com.waffiq.bazz_movies.feature.detail.domain.usecase.getDetailMovie
 
 import androidx.paging.PagingData
+import com.waffiq.bazz_movies.core.domain.ResultItem
 import com.waffiq.bazz_movies.core.network.utils.result.NetworkResult
 import com.waffiq.bazz_movies.core.utils.GenreHelper.transformListGenreToJoinString
 import com.waffiq.bazz_movies.core.utils.GenreHelper.transformToGenreIDs
@@ -61,6 +62,6 @@ class GetDetailMovieInteractor @Inject constructor(
     detailRepository.getCreditMovies(movieId)
 
   override fun getPagingMovieRecommendation(movieId: Int)
-    : Flow<PagingData<com.waffiq.bazz_movies.core.data.ResultItem>> =
+    : Flow<PagingData<ResultItem>> =
     detailRepository.getPagingMovieRecommendation(movieId)
 }
