+++ feature/detail/src/main/kotlin/com/waffiq/bazz_movies/feature/detail/domain/usecase/getDetailTv/GetDetailTvUseCase.kt
@@ -1,6 +1,7 @@
 package com.waffiq.bazz_movies.feature.detail.domain.usecase.getDetailTv
 
 import androidx.paging.PagingData
+import com.waffiq.bazz_movies.core.domain.ResultItem
 import com.waffiq.bazz_movies.core.network.utils.result.NetworkResult
 import com.waffiq.bazz_movies.feature.detail.domain.model.DetailMovieTvUsed
 import com.waffiq.bazz_movies.feature.detail.domain.model.MovieTvCredits
@@ -12,5 +13,5 @@ interface GetDetailTvUseCase {
   suspend fun getExternalTvId(tvId: Int): Flow<NetworkResult<ExternalTvID>>
   suspend fun getTrailerLinkTv(tvId: Int): Flow<NetworkResult<String>>
   suspend fun getCreditTv(tvId: Int): Flow<NetworkResult<MovieTvCredits>>
-  fun getPagingTvRecommendation(tvId: Int): Flow<PagingData<com.waffiq.bazz_movies.core.data.ResultItem>>
+  fun getPagingTvRecommendation(tvId: Int): Flow<PagingData<ResultItem>>
 }
