+++ feature/favorite/src/main/kotlin/com/waffiq/bazz_movies/feature/favorite/domain/usecase/GetFavoriteMovieUseCase.kt
@@ -1,6 +1,6 @@
 package com.waffiq.bazz_movies.feature.favorite.domain.usecase
 
-import com.waffiq.bazz_movies.core.data.ResultItem
+import com.waffiq.bazz_movies.core.domain.ResultItem
 import kotlinx.coroutines.flow.Flow
 
 interface GetFavoriteMovieUseCase {
