+++ feature/favorite/src/main/kotlin/com/waffiq/bazz_movies/feature/favorite/domain/usecase/GetFavoriteTvInteractor.kt
@@ -1,6 +1,6 @@
 package com.waffiq.bazz_movies.feature.favorite.domain.usecase
 
-import com.waffiq.bazz_movies.core.data.ResultItem
+import com.waffiq.bazz_movies.core.domain.ResultItem
 import com.waffiq.bazz_movies.core.movie.domain.repository.IMoviesRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
