+++ feature/home/src/main/kotlin/com/waffiq/bazz_movies/feature/home/domain/repository/IHomeRepository.kt
@@ -1,7 +1,7 @@
 package com.waffiq.bazz_movies.feature.home.domain.repository
 
 import androidx.paging.PagingData
-import com.waffiq.bazz_movies.core.data.ResultItem
+import com.waffiq.bazz_movies.core.domain.ResultItem
 import kotlinx.coroutines.flow.Flow
 
 interface IHomeRepository {
