+++ navigation/build.gradle.kts
@@ -6,7 +6,6 @@ plugins {
 android.namespace = "com.waffiq.bazz_movies.navigation"
 
 dependencies {
-  implementation(project(":core:data"))
   implementation(project(":core:domain"))
 }
