+++ navigation/src/main/kotlin/com/waffiq/bazz_movies/navigation/INavigator.kt
@@ -2,7 +2,7 @@ package com.waffiq.bazz_movies.navigation
 
 import android.app.Activity
 import android.content.Context
-import com.waffiq.bazz_movies.core.data.ResultItem
+import com.waffiq.bazz_movies.core.domain.ResultItem
 import com.waffiq.bazz_movies.core.domain.MovieTvCastItem
 
 interface INavigator {
